+++ src/main/java/org/telegram/bot/interpreterbot/repository/ClientCacheRepository.java
@@ -15,31 +15,31 @@ public class ClientCacheRepository {
     @Value("${spring.redis.key.client}")
     private String clientKey;
 
-    private final RedisTemplate<Integer, Client> redisTemplate;
+    private final RedisTemplate<Long, Client> redisTemplate;
 
     private final HashOperations hashOperations;
 
-    public ClientCacheRepository(RedisTemplate<Integer, Client> redisTemplate) {
+    public ClientCacheRepository(RedisTemplate<Long, Client> redisTemplate) {
         this.redisTemplate = redisTemplate;
         hashOperations = redisTemplate.opsForHash();
     }
 
-    public void save(Integer clientId, Client client) {
+    public void save(Long clientId, Client client) {
         hashOperations.put(clientKey, clientId, client);
     }
 
-    public Client findById(Integer clientId) {
+    public Client findById(Long clientId) {
         return Optional.ofNullable((Client) hashOperations.get(clientKey, clientId))
                 .orElseThrow(NoSuchElementException::new);
     }
 
-    public Client findByIdAndAuthTrue(Integer clientId) {
+    public Client findByIdAndAuthTrue(Long clientId) {
         return Optional.ofNullable((Client) hashOperations.get(clientKey, clientId))
                 .filter(Client::isAuth)
                 .orElseThrow(NoSuchElementException::new);
     }
 
-    public void delete(Integer clientId) {
+    public void delete(Long clientId) {
         hashOperations.delete(clientKey, clientId);
     }
