+++ src/main/java/org/telegram/bot/interpreterbot/service/ClientService.java
@@ -6,5 +6,5 @@ public interface ClientService {
 
     void save(Client client);
 
-    Client findById(int id);
+    Client findById(Long id);
 }
